Refactor component registry imports and add registry path aliases
