+++ src/features/editor/hooks/useAllComponentMetas.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { componentsRegistry } from '@src/lib/componentRegistry'
+import { componentsRegistry } from '@registry/componentRegistry'
 import type { ComponentMeta } from '@components/site/types'
 
 export function useAllComponentMetas() {
