+++ src/features/editor/store/usePageStore.ts
@@ -2,7 +2,7 @@ import { create } from 'zustand'
 import { v4 as uuidv4 } from 'uuid'
 import type { PageStore } from '@features/editor/types'
 import type { ComponentMeta } from '@components/site/types'
-import { componentsRegistry } from '@src/lib/componentRegistry'
+import { componentsRegistry } from '@registry/componentRegistry'
 import { extractDefaultProps } from '@src/features/editor/utils/extractDefaultProps'
 
 export const usePageStore = create<PageStore>((set) => ({
