+++ src/registry/componentRegistry.ts
