+++ tsconfig.app.json
@@ -36,6 +36,7 @@
       "@src/*": ["src/*"],
       "@types/*": ["src/types/*"],
       "@lib/*": ["src/lib/*"],
+      "@registry/*": ["src/registry/*"],
       "@constants/*": ["src/constants/*"]
     }
   },
